// kiesmbo.d.ts
export interface paths {
    "/api/v1/export": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Export */
        get: operations["Export"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/studieincijfers/{brin}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** LinksForBRIN */
        get: operations["LinksForBRIN"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/studieincijfers/{brin}/{code}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** LeafletForCode */
        get: operations["LeafletForCode"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/studies": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Studies */
        get: operations["Studies"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/studies/{studyNumber}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** StudyDetails */
        get: operations["StudyDetails"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v2/export": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** ExportV2 */
        get: operations["ExportV2"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v2/studiesexport": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** StudiesexportV2 */
        get: operations["StudiesexportV2"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        fullExport: {
            /** Format: date-time */
            lastUpdated?: string;
            schools?: components["schemas"]["school"][];
            studies?: components["schemas"]["study"][];
        };
        fullExportV2: {
            /** Format: date-time */
            lastUpdated?: string;
            schools?: components["schemas"]["school"][];
            studies?: components["schemas"]["studyV2"][];
        };
        school: {
            brin?: string;
            name?: string;
            website?: string;
            logoUrl?: string
            locations?: components["schemas"]["location"][];
        };
        study: {
            crebo?: string;
            creboName?: string;
            studyNumber?: string;
            profession?: string;
            professionNumber?: string;
            /** Format: int32 */
            level?: number;
            duration?: string;
            introtext?: string;
            youtubeVideoCode?: string;
        };
        studyV2: {
            crebo?: string;
            creboName?: string;
            studyNumber?: string;
            profession?: string;
            professionNumber?: string;
            /** Format: int32 */
            level?: number;
            duration?: string;
            introtext?: string;
            youtubeVideoCode?: string;
            educationContent?: string;
        };
        location: {
            brinvest?: string;
            name?: string;
            isMainLocation?: boolean;
            street?: string;
            houseNumber?: string;
            zipCode?: string;
            city?: string;
            province?: string;
            website?: string;
            "geoCoordinate": {
                "longitude": number,
                "latitude": number
            }
            studies?: components["schemas"]["locationStudy"][];
            opendays?: components["schemas"]["locationOpenday"][];
        };
        locationStudy: {
            crebo?: string;
            learningPaths?: string[];
            startingFebruary?: boolean;
            startingSeptember?: boolean;
            startingAnytime?: boolean;
            hasNumerusFixus?: boolean;
            urls?: components["schemas"]["studyUrl"][];
        };
        locationOpenday: {
            /** Format: date-time */
            startDate?: string;
            /** Format: date-time */
            endDate?: string;
            url?: string;
            type?: string;
            isOnline?: boolean;
            isGeneralOpenDay?: boolean;
        };
        studyUrl: {
            title?: string;
            url?: string;
        };
        studiesExportV2: {
            /** Format: date-time */
            lastUpdated?: string;
            studies?: components["schemas"]["studyV2"][];
        };
        apiResult: {
            hasError?: boolean;
            errorMessage?: string;
            data?: Record<string, never>;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
    Export: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    lastUpdated: components["schemas"]["fullExport"];
                    schools: components["schemas"]["fullExport"];
                    studies: components["schemas"]["fullExport"];
                };
            };
        };
    };
    LinksForBRIN: {
        parameters: {
            query?: never;
            header?: never;
            path: {
                brin: string;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    HasError: components["schemas"]["apiResult"];
                    Data: components["schemas"]["apiResult"];
                };
            };
        };
    };
    LeafletForCode: {
        parameters: {
            query?: never;
            header?: never;
            path: {
                brin: string;
                code: string;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    HasError: components["schemas"]["apiResult"];
                    Data: components["schemas"]["apiResult"];
                };
            };
        };
    };
    Studies: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    HasError: components["schemas"]["apiResult"];
                    Data: components["schemas"]["apiResult"];
                };
            };
        };
    };
    StudyDetails: {
        parameters: {
            query?: never;
            header?: never;
            path: {
                studyNumber: string;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    HasError: components["schemas"]["apiResult"];
                    Data: components["schemas"]["apiResult"];
                };
            };
        };
    };
    ExportV2: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    lastUpdated: components["schemas"]["fullExportV2"];
                    schools: components["schemas"]["fullExportV2"];
                    studies: components["schemas"]["fullExportV2"];
                };
            };
        };
    };
    StudiesexportV2: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    lastUpdated: components["schemas"]["studiesExportV2"];
                    studies: components["schemas"]["studiesExportV2"];
                };
            };
        };
    };
}